requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
enum34==1.1.6
google-api-python-client==1.6.2
httplib2==0.10.3
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.9.1
MarkupSafe==0.23
oauth2client==4.1.0
packaging==16.8
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
Pygments==2.1.3
pyparsing==2.2.0
PyYAML==3.12
requests==2.13.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
uritemplate==3.0.0
virtualenv==15.0.3
wcwidth==0.1.7
rope
jedi
flake8
importmagic
autopep8
yapf
